# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ebdaa Skill Matrix System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /skills:
    get:
      operationId: listSkills
      tags: [skills]
      summary: List skills with optional filters
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: department_id
          in: query
          schema:
            type: string
        - name: criticality
          in: query
          schema:
            type: string
            enum: [Low, Medium, High, Critical]
        - name: is_active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
    post:
      operationId: createSkill
      tags: [skills]
      summary: Create skill
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkillBody"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
  /skills/{id}:
    get:
      operationId: getSkill
      tags: [skills]
      summary: Get skill by ID
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"
    put:
      operationId: updateSkill
      tags: [skills]
      summary: Update skill
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSkillBody"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Skill"

  # ── CAMPAIGNS ─────────────────────────────────────────────────────────────────
...
  /training:
    get:
      operationId: listTrainingRecommendations
      tags: [training]
      summary: List training recommendations
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: employee_id
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [Pending, "In Progress", Completed, Cancelled]
        - name: recommendation_type
          in: query
          schema:
            type: string
            enum: [Immediate, Short-term, Long-term, Promotion]
      responses:
        "200":
          description: Training recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrainingRecommendation"
    post:
      operationId: createTrainingRecommendation
      tags: [training]
      summary: Create training recommendation
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingBody"
      responses:
        "201":
          description: Created recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRecommendation"
  /training/{id}:
    put:
      operationId: updateTrainingRecommendation
      tags: [training]
      summary: Update training recommendation status
      parameters:
        - name: x-user-token
          in: header
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrainingBody"
      responses:
        "200":
          description: Updated recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingRecommendation"
components:
  schemas:
...
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        manager_id:
          type: string
          nullable: true
        manager_email:
          type: string
          nullable: true
        employee_count:
          type: integer
        created_at:
          type: string
      required: [id, name, employee_count, created_at]
...
    Skill:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
          nullable: true
        name:
          type: string
        department_id:
          type: string
        category:
          type: string
          nullable: true
        weight:
          type: integer
        criticality:
          type: string
          enum: [Low, Medium, High, Critical]
        description:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
        department:
          $ref: "#/components/schemas/Department"
      required: [id, name, department_id, weight, criticality, is_active, created_at, department]
    CreateSkillBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        department_id:
          type: string
        category:
          type: string
        weight:
          type: integer
          minimum: 1
          maximum: 5
        criticality:
          type: string
          enum: [Low, Medium, High, Critical]
        description:
          type: string
      required: [name, department_id, weight, criticality]

    # ── CAMPAIGNS ─────────────────────────────────────────────────────────────
...
    TrainingRecommendation:
      type: object
      properties:
        id:
          type: string
        employee_id:
          type: string
        skill_id:
          type: string
          nullable: true
        campaign_id:
          type: string
          nullable: true
        score:
          type: integer
          nullable: true
        recommendation_type:
          type: string
          enum: [Immediate, Short-term, Long-term, Promotion]
        status:
          type: string
          enum: [Pending, "In Progress", Completed, Cancelled]
        assigned_trainer_id:
          type: string
          nullable: true
        target_date:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        employee_name:
          type: string
          nullable: true
        skill_name:
          type: string
          nullable: true
      required: [id, employee_id, recommendation_type, status, created_at]
    CreateTrainingBody:
      type: object
      properties:
        employee_id:
          type: string
        skill_id:
          type: string
        campaign_id:
          type: string
        score:
          type: integer
        recommendation_type:
          type: string
          enum: [Immediate, Short-term, Long-term, Promotion]
        target_date:
          type: string
        notes:
          type: string
      required: [employee_id, recommendation_type]
    UpdateTrainingBody:
      type: object
      properties:
        status:
          type: string
          enum: [Pending, "In Progress", Completed, Cancelled]
        assigned_trainer_id:
          type: string
        target_date:
          type: string
        notes:
          type: string

    # ── DASHBOARD ─────────────────────────────────────────────────────────────
...